consildate everything into just converse_itnernal and convcore
